Consolidate identifier metadata refresh tests into Core test classes
