+++ infra/common/src/test/java/org/apache/shardingsphere/infra/metadata/database/schema/model/ShardingSphereSchemaTest.java
@@ -17,11 +17,18 @@
 
 package org.apache.shardingsphere.infra.metadata.database.schema.model;
 
+import org.apache.shardingsphere.database.connector.core.metadata.identifier.IdentifierCasePolicy;
+import org.apache.shardingsphere.database.connector.core.metadata.identifier.IdentifierCasePolicyFactory;
+import org.apache.shardingsphere.database.connector.core.metadata.identifier.IdentifierCasePolicySet;
+import org.apache.shardingsphere.database.connector.core.metadata.identifier.IdentifierScope;
 import org.apache.shardingsphere.database.connector.core.type.DatabaseType;
+import org.apache.shardingsphere.infra.metadata.identifier.DatabaseIdentifierContext;
 import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
 import org.junit.jupiter.api.Test;
 
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.EnumMap;
 import java.util.HashSet;
 import java.util.Map;
 
@@ -33,30 +40,32 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class ShardingSphereSchemaTest {
     
-    private final DatabaseType databaseType = TypedSPILoader.getService(DatabaseType.class, "PostgreSQL");
+    private final DatabaseType postgreSQLDatabaseType = TypedSPILoader.getService(DatabaseType.class, "PostgreSQL");
+    
+    private final DatabaseType oracleDatabaseType = TypedSPILoader.getService(DatabaseType.class, "Oracle");
     
     @Test
     void assertGetAllTables() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-        assertThat(new HashSet<>(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList())
+        assertThat(new HashSet<>(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList())
                 .getAllTables()), is(Collections.singleton(table)));
     }
     
     @Test
     void assertContainsTable() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-        assertTrue(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).containsTable("foo_tbl"));
+        assertTrue(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).containsTable("foo_tbl"));
     }
     
     @Test
     void assertGetTable() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-        assertThat(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).getTable("foo_tbl"), is(table));
+        assertThat(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).getTable("foo_tbl"), is(table));
     }
     
     @Test
     void assertPutTable() {
-        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", databaseType);
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", postgreSQLDatabaseType);
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
         schema.putTable(table);
         assertThat(schema.getTable("foo_tbl"), is(table));
@@ -65,7 +74,7 @@ class ShardingSphereSchemaTest {
     @Test
     void assertRemoveTable() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
         schema.removeTable("foo_tbl");
         assertNull(schema.getTable("foo_tbl"));
     }
@@ -73,32 +82,32 @@ class ShardingSphereSchemaTest {
     @Test
     void assertGetAllViews() {
         ShardingSphereView view = new ShardingSphereView("foo_view", "SELECT 1");
-        assertThat(new HashSet<>(new ShardingSphereSchema("foo_db", databaseType, Collections.emptyList(), Collections.singleton(view))
+        assertThat(new HashSet<>(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view))
                 .getAllViews()), is(Collections.singleton(view)));
     }
     
     @Test
     void assertContainsView() {
         ShardingSphereView view = new ShardingSphereView("foo_view", "SELECT 1");
-        assertTrue(new ShardingSphereSchema("foo_db", databaseType, Collections.emptyList(), Collections.singleton(view)).containsView("foo_view"));
+        assertTrue(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view)).containsView("foo_view"));
     }
     
     @Test
     void assertGetView() {
         ShardingSphereView view = new ShardingSphereView("foo_view", "SELECT 1");
-        assertThat(new ShardingSphereSchema("foo_db", databaseType, Collections.emptyList(), Collections.singleton(view)).getView("foo_view"), is(view));
+        assertThat(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view)).getView("foo_view"), is(view));
     }
     
     @Test
     void assertPutView() {
-        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", databaseType, Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.emptyList(), Collections.emptyList());
         schema.putView(new ShardingSphereView("foo_view", "SELECT * FROM test_table"));
         assertTrue(schema.containsView("foo_view"));
     }
     
     @Test
     void assertRemoveView() {
-        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", databaseType, Collections.emptyList(),
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.emptyList(),
                 Collections.singleton(new ShardingSphereView("foo_view", "SELECT * FROM test_table")));
         schema.removeView("foo_view");
         assertFalse(schema.containsView("foo_view"));
@@ -108,45 +117,46 @@ class ShardingSphereSchemaTest {
     void assertContainsIndex() {
         ShardingSphereTable table = new ShardingSphereTable(
                 "foo_tbl", Collections.emptyList(), Collections.singleton(new ShardingSphereIndex("col_idx", Collections.emptyList(), false)), Collections.emptyList());
-        assertTrue(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).containsIndex("foo_tbl", "col_idx"));
+        assertTrue(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).containsIndex("foo_tbl", "col_idx"));
     }
     
     @Test
     void assertContainsIndexWithIndexNotExists() {
         ShardingSphereTable table = new ShardingSphereTable(
                 "foo_tbl", Collections.emptyList(), Collections.singleton(new ShardingSphereIndex("col_idx", Collections.emptyList(), false)), Collections.emptyList());
-        assertFalse(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).containsIndex("foo_tbl", "foo_idx"));
+        assertFalse(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).containsIndex("foo_tbl", "foo_idx"));
     }
     
     @Test
     void assertContainsIndexWithTableNotExists() {
-        assertFalse(new ShardingSphereSchema("foo_db", databaseType).containsIndex("nonexistent_tbl", "nonexistent_idx"));
+        assertFalse(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType).containsIndex("nonexistent_tbl", "nonexistent_idx"));
     }
     
     @Test
     void assertGetVisibleColumnNamesWhenTableNotExists() {
-        assertTrue(new ShardingSphereSchema("foo_tbl", databaseType, Collections.emptyList(), Collections.emptyList()).getVisibleColumnNames("nonexistent_tbl").isEmpty());
+        assertTrue(new ShardingSphereSchema("foo_tbl", postgreSQLDatabaseType, Collections.emptyList(), Collections.emptyList()).getVisibleColumnNames("nonexistent_tbl").isEmpty());
     }
     
     @Test
     void assertGetVisibleColumnNamesWhenContainsKey() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singletonList(
                 new ShardingSphereColumn("foo_col", 0, false, false, false, true, false, false)), Collections.emptyList(), Collections.emptyList());
-        assertThat(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).getVisibleColumnNames("foo_tbl"), is(Collections.singletonList("foo_col")));
+        assertThat(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).getVisibleColumnNames("foo_tbl"),
+                is(Collections.singletonList("foo_col")));
     }
     
     @Test
     void assertGetVisibleColumnNamesWhenNotContainsKey() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singletonList(
                 new ShardingSphereColumn("foo_col", 0, false, false, false, false, true, false)), Collections.emptyList(), Collections.emptyList());
-        assertTrue(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).getVisibleColumnNames("foo_tbl").isEmpty());
+        assertTrue(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).getVisibleColumnNames("foo_tbl").isEmpty());
     }
     
     @Test
     void assertGetVisibleColumnAndIndexMapWhenContainsTable() {
         ShardingSphereColumn column = new ShardingSphereColumn("foo_col", 0, false, false, false, true, false, false);
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singletonList(column), Collections.emptyList(), Collections.emptyList());
-        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
         Map<String, Integer> actual = schema.getVisibleColumnAndIndexMap("foo_tbl");
         assertThat(actual.size(), is(1));
         assertTrue(actual.containsKey("foo_col"));
@@ -154,23 +164,160 @@ class ShardingSphereSchemaTest {
     
     @Test
     void assertGetVisibleColumnAndIndexMapWhenNotContainsTable() {
-        assertTrue(new ShardingSphereSchema("foo_db", databaseType).getVisibleColumnAndIndexMap("nonexistent_tbl").isEmpty());
+        assertTrue(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType).getVisibleColumnAndIndexMap("nonexistent_tbl").isEmpty());
     }
     
     @Test
     void assertIsEmptyWithEmptyTable() {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
-        assertFalse(new ShardingSphereSchema("foo_db", databaseType, Collections.singleton(table), Collections.emptyList()).isEmpty());
+        assertFalse(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList()).isEmpty());
     }
     
     @Test
     void assertIsEmptyWithEmptyView() {
         ShardingSphereView view = new ShardingSphereView("foo_view", "");
-        assertFalse(new ShardingSphereSchema("foo_db", databaseType, Collections.emptyList(), Collections.singleton(view)).isEmpty());
+        assertFalse(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view)).isEmpty());
     }
     
     @Test
     void assertIsEmpty() {
-        assertTrue(new ShardingSphereSchema("foo_db", databaseType).isEmpty());
+        assertTrue(new ShardingSphereSchema("foo_db", postgreSQLDatabaseType).isEmpty());
+    }
+    
+    @Test
+    void assertContainsUpperCaseTable() {
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertTrue(schema.containsTable("FOO_TBL"));
+    }
+    
+    @Test
+    void assertContainsTableWithOracleRule() {
+        ShardingSphereTable table = new ShardingSphereTable("FOO_TBL", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", oracleDatabaseType, Collections.singleton(table), Collections.emptyList());
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newUpperCasePolicySet()));
+        assertTrue(schema.containsTable("foo_tbl"));
+    }
+    
+    @Test
+    void assertGetUpperCaseTable() {
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertThat(schema.getTable("FOO_TBL"), is(table));
+    }
+    
+    @Test
+    void assertContainsUpperCaseView() {
+        ShardingSphereView view = new ShardingSphereView("foo_view", "SELECT 1");
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view));
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertTrue(schema.containsView("FOO_VIEW"));
+    }
+    
+    @Test
+    void assertContainsViewWithOracleRule() {
+        ShardingSphereView view = new ShardingSphereView("FOO_VIEW", "SELECT 1");
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", oracleDatabaseType, Collections.emptyList(), Collections.singleton(view));
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newUpperCasePolicySet()));
+        assertTrue(schema.containsView("foo_view"));
+    }
+    
+    @Test
+    void assertGetUpperCaseView() {
+        ShardingSphereView view = new ShardingSphereView("foo_view", "SELECT 1");
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view));
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertThat(schema.getView("FOO_VIEW"), is(view));
+    }
+    
+    @Test
+    void assertAttachIdentifierContext() {
+        ShardingSphereTable table = new ShardingSphereTable("Foo_Tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        assertTrue(schema.containsTable("FOO_TBL"));
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertFalse(schema.containsTable("FOO_TBL"));
+    }
+    
+    @Test
+    void assertAttachIdentifierContextToTable() {
+        ShardingSphereColumn column = new ShardingSphereColumn("Foo_Col", java.sql.Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singleton(column), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertFalse(schema.getTable("foo_tbl").containsColumn("FOO_COL"));
+    }
+    
+    @Test
+    void assertContainsTableByLogicalTableIndexWhenHeterogeneousLookupEnabled() {
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        Map<IdentifierScope, IdentifierCasePolicy> scopedRules = new EnumMap<>(IdentifierScope.class);
+        scopedRules.put(IdentifierScope.LOGICAL_TABLE, IdentifierCasePolicyFactory.newLowerCasePolicySet().getPolicy(IdentifierScope.TABLE));
+        DatabaseIdentifierContext context = new DatabaseIdentifierContext(new IdentifierCasePolicySet(IdentifierCasePolicyFactory.newUpperCasePolicySet()
+                .getPolicy(IdentifierScope.TABLE), scopedRules), true);
+        schema.refreshIdentifierContext(context);
+        assertTrue(schema.containsTable("FOO_TBL"));
+        assertThat(schema.getTable("FOO_TBL"), is(table));
+    }
+    
+    @Test
+    void assertGetTablePrioritizesPhysicalTableIndexWhenBothIndexesCanMatch() {
+        ShardingSphereTable lowerCaseTable = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereTable upperCaseTable = new ShardingSphereTable("FOO_TBL", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Arrays.asList(lowerCaseTable, upperCaseTable), Collections.emptyList());
+        Map<IdentifierScope, IdentifierCasePolicy> scopedRules = new EnumMap<>(IdentifierScope.class);
+        scopedRules.put(IdentifierScope.LOGICAL_TABLE, IdentifierCasePolicyFactory.newLowerCasePolicySet().getPolicy(IdentifierScope.TABLE));
+        DatabaseIdentifierContext context = new DatabaseIdentifierContext(new IdentifierCasePolicySet(IdentifierCasePolicyFactory.newUpperCasePolicySet()
+                .getPolicy(IdentifierScope.TABLE), scopedRules), true);
+        schema.refreshIdentifierContext(context);
+        assertThat(schema.getTable("FOO_TBL"), is(upperCaseTable));
+    }
+    
+    @Test
+    void assertRemoveTableByLogicalTableIndexWhenHeterogeneousLookupEnabled() {
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        Map<IdentifierScope, IdentifierCasePolicy> scopedRules = new EnumMap<>(IdentifierScope.class);
+        scopedRules.put(IdentifierScope.LOGICAL_TABLE, IdentifierCasePolicyFactory.newLowerCasePolicySet().getPolicy(IdentifierScope.TABLE));
+        DatabaseIdentifierContext context = new DatabaseIdentifierContext(new IdentifierCasePolicySet(IdentifierCasePolicyFactory.newUpperCasePolicySet()
+                .getPolicy(IdentifierScope.TABLE), scopedRules), true);
+        schema.refreshIdentifierContext(context);
+        schema.removeTable("FOO_TBL");
+        assertNull(schema.getTable("foo_tbl"));
+    }
+    
+    @Test
+    void assertRemoveTableByActualName() {
+        ShardingSphereTable table = new ShardingSphereTable("Foo_Tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.singleton(table), Collections.emptyList());
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        schema.removeTable("Foo_Tbl");
+        assertTrue(schema.getAllTables().isEmpty());
+    }
+    
+    @Test
+    void assertRemoveTablePrioritizesPhysicalTableIndexWhenBothIndexesCanMatch() {
+        ShardingSphereTable lowerCaseTable = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereTable upperCaseTable = new ShardingSphereTable("FOO_TBL", Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Arrays.asList(lowerCaseTable, upperCaseTable), Collections.emptyList());
+        Map<IdentifierScope, IdentifierCasePolicy> scopedRules = new EnumMap<>(IdentifierScope.class);
+        scopedRules.put(IdentifierScope.LOGICAL_TABLE, IdentifierCasePolicyFactory.newLowerCasePolicySet().getPolicy(IdentifierScope.TABLE));
+        DatabaseIdentifierContext context = new DatabaseIdentifierContext(new IdentifierCasePolicySet(IdentifierCasePolicyFactory.newUpperCasePolicySet()
+                .getPolicy(IdentifierScope.TABLE), scopedRules), true);
+        schema.refreshIdentifierContext(context);
+        schema.removeTable("FOO_TBL");
+        assertThat(schema.getTable("FOO_TBL"), is(lowerCaseTable));
+    }
+    
+    @Test
+    void assertRemoveViewByActualName() {
+        ShardingSphereView view = new ShardingSphereView("Foo_View", "SELECT 1");
+        ShardingSphereSchema schema = new ShardingSphereSchema("foo_schema", postgreSQLDatabaseType, Collections.emptyList(), Collections.singleton(view));
+        schema.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        schema.removeView("Foo_View");
+        assertTrue(schema.getAllViews().isEmpty());
     }
 }
