+++ infra/common/src/test/java/org/apache/shardingsphere/infra/metadata/database/schema/model/ShardingSphereTableTest.java
@@ -17,6 +17,8 @@
 
 package org.apache.shardingsphere.infra.metadata.database.schema.model;
 
+import org.apache.shardingsphere.database.connector.core.metadata.identifier.IdentifierCasePolicyFactory;
+import org.apache.shardingsphere.infra.metadata.identifier.DatabaseIdentifierContext;
 import org.junit.jupiter.api.Test;
 
 import java.sql.Types;
@@ -134,5 +136,63 @@ class ShardingSphereTableTest {
         ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.singleton(constraint));
         assertThat(table.getAllConstraints().size(), is(1));
         assertThat(table.getAllConstraints(), hasItems(constraint));
+    }
+    
+    @Test
+    void assertContainsUpperCaseColumn() {
+        ShardingSphereColumn column = new ShardingSphereColumn("foo_col", Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singleton(column), Collections.emptyList(), Collections.emptyList());
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertTrue(table.containsColumn("FOO_COL"));
+    }
+    
+    @Test
+    void assertContainsColumnWithOracleRule() {
+        ShardingSphereColumn column = new ShardingSphereColumn("FOO_COL", Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singleton(column), Collections.emptyList(), Collections.emptyList());
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newUpperCasePolicySet()));
+        assertTrue(table.containsColumn("foo_col"));
+    }
+    
+    @Test
+    void assertGetUpperCaseColumn() {
+        ShardingSphereColumn column = new ShardingSphereColumn("foo_col", Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singleton(column), Collections.emptyList(), Collections.emptyList());
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertThat(table.getColumn("FOO_COL"), is(column));
+    }
+    
+    @Test
+    void assertFindColumnNamesIfNotExistedFrom() {
+        ShardingSphereColumn column1 = new ShardingSphereColumn("foo_col", Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereColumn column2 = new ShardingSphereColumn("bar_col", Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Arrays.asList(column1, column2), Collections.emptyList(), Collections.emptyList());
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertThat(table.findColumnNamesIfNotExistedFrom(Collections.singleton("FOO_COL")), is(Collections.singleton("bar_col")));
+    }
+    
+    @Test
+    void assertGetIndex() {
+        ShardingSphereIndex index = new ShardingSphereIndex("foo_idx", Collections.emptyList(), false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.singleton(index), Collections.emptyList());
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertThat(table.getIndex("FOO_IDX"), is(index));
+    }
+    
+    @Test
+    void assertContainsConstraint() {
+        ShardingSphereConstraint constraint = new ShardingSphereConstraint("FOO_FK", "ref_tbl");
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.emptyList(), Collections.emptyList(), Collections.singleton(constraint));
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newUpperCasePolicySet()));
+        assertTrue(table.containsConstraint("foo_fk"));
+    }
+    
+    @Test
+    void assertRefreshIdentifierContext() {
+        ShardingSphereColumn column = new ShardingSphereColumn("Foo_Col", Types.INTEGER, false, true, false, true, false, false);
+        ShardingSphereTable table = new ShardingSphereTable("foo_tbl", Collections.singleton(column), Collections.emptyList(), Collections.emptyList());
+        assertTrue(table.containsColumn("FOO_COL"));
+        table.refreshIdentifierContext(new DatabaseIdentifierContext(IdentifierCasePolicyFactory.newLowerCasePolicySet()));
+        assertFalse(table.containsColumn("FOO_COL"));
     }
 }
